gai-inpainting-cdk/notebooks/code/requirements.txt
nvgpu==0.9.0
pynvml==11.4.1
opencv-python
torch
torchvision
Pillow
transformers
numpy
matplotlib
diffusers
albumentations
accelerate